+++ network/lorawan_4_4_2/lora/mac/region/RegionUS915.c
@@ -504,7 +504,7 @@ bool RegionUS915AdrNext( AdrNextParams_t* adrNext, int8_t* drOut, int8_t* txPowO
 
 void RegionUS915ComputeRxWindowParameters( int8_t datarate, uint8_t minRxSymbols, uint32_t rxError, RxConfigParams_t *rxConfigParams )
 {
-    double tSymbol = 0.0;
+    DECIMAL tSymbol = 0.0;
     uint32_t radioWakeUpTime;
 
     rxConfigParams->Datarate = datarate;
